autogen/requirements2.txt
pip-system-certs

ag2[openai]
chromadb
sentence_transformers
numpy

mcp
fastmcp
autogen-ext[openai,mcp]
autogen_agentchat

asyncio
python-dotenv

fastapi
uvicorn
jinja2